*Add captions
Javascript file now gets the URL of the images and their captions from
an array in the JSON file. These are displayed on top of the image using
CSS absolute positioning and opacity is reduced.

*Change font and layout

Changed the font and experimented with different padding, margin and the
use of flexbox justify content to keep the grid of images neat when
screen width changes.